Refactor to group by function not aggregate
