+++ examples/postgres_payments/src/aggregate/mod.rs
@@ -1,0 +1,2 @@
+pub mod account_aggregate;
+pub mod credit_aggregate;
