--- examples/postgres_payments/src/bank_account/mod.rs
@@ -1,6 +1,0 @@
-pub mod aggregate;
-pub mod command;
-pub mod error;
-pub mod event;
-pub mod projector;
-pub mod state;
