--- examples/postgres_payments/src/credit_card/command.rs
@@ -1,4 +1,0 @@
-pub enum CreditCardCommand {
-    Pay { amount: i32 },
-    Refund { amount: i32 },
-}
