--- examples/postgres_payments/src/credit_card/event.rs
@@ -1,7 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Debug)]
-pub enum CreditCardEvent {
-    Payed { amount: i32 },
-    Refunded { amount: i32 },
-}
