--- examples/postgres_payments/src/credit_card/mod.rs
@@ -1,7 +1,0 @@
-pub mod aggregate;
-pub mod command;
-pub mod error;
-pub mod event;
-pub mod policy;
-pub mod projector;
-pub mod state;
