+++ examples/postgres_payments/src/event.rs
@@ -5,3 +5,9 @@ pub enum BankAccountEvent {
     Withdrawn { amount: i32 },
     Deposited { amount: i32 },
 }
+
+#[derive(Serialize, Deserialize, Debug)]
+pub enum CreditCardEvent {
+    Payed { amount: i32 },
+    Refunded { amount: i32 },
+}
