+++ examples/postgres_payments/src/lib.rs
@@ -1,2 +1,7 @@
-pub mod bank_account;
-pub mod credit_card;
+pub mod aggregate;
+pub mod projector;
+pub mod state;
+pub mod error;
+pub mod command;
+pub mod event;
+pub mod policy;
