+++ examples/postgres_payments/src/projector/mod.rs
@@ -1,0 +1,2 @@
+pub mod account_projector;
+pub mod credit_projector;
