+++ examples/postgres_payments/tests/delete_bank_account_test.rs
@@ -3,10 +3,10 @@ use sqlx::{Pool, Postgres};
 use uuid::Uuid;
 
 use esrs::aggregate::{AggregateManager, AggregateState, Eraser};
-use postgres_payments::bank_account::aggregate::BankAccountAggregate;
-use postgres_payments::bank_account::command::BankAccountCommand;
-use postgres_payments::bank_account::projector::BankAccount;
-use postgres_payments::bank_account::state::BankAccountState;
+use postgres_payments::aggregate::account_aggregate::BankAccountAggregate;
+use postgres_payments::command::BankAccountCommand;
+use postgres_payments::projector::account_projector::BankAccount;
+use postgres_payments::state::account_state::BankAccountState;
 
 #[tokio::test]
 async fn postgres_delete_bank_account_aggregate_and_read_model_test() {
